Refactor Sidebar, moving hooks to a separate file
